requirements.txt
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.19
beautifulsoup4==4.12.2
faiss-cpu==1.7.4
tiktoken==0.5.2
duckduckgo-search==3.9.6
python-dotenv==1.0.0
openai==1.3.0
requests==2.31.0
streamlit==1.28.0
pypdf==3.17.0